pyproject.toml
[project]
name = "real-time-timescaledb"
version = "1.0.0"
description = "This repository contains the code for the blog post 'Working with Real-Time Data? TimescaleDB Might Be the Simple, Scalable Solution You Need'"
authors = [
    {name = "Jag Reehal", email = "jag@jagreehal.com"},
]
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "psycopg2-binary==2.9.9",
    "sqlalchemy==2.0.23",
    "sqlalchemy-timescaledb==0.4.1",
    "python-dotenv==1.0.0",
    "prettytable>=3.12.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=7.4.0",
    "mypy>=1.5.0",
    "black>=23.9.1",
    "uv>=0.2.0",
    "ipykernel>=6.0.0"
]

[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

# Fix per-file-ignores using table format
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["python_template"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "src",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.black]
line-length = 120
target-version = ['py311']
